Add support for non-int times

+++ python/games/board.py
@@ -109,8 +109,8 @@ class MessageBoard(game.Game):
 
     def add_message(self):
         text = request.form.get("text", "?")
-        priority = int(request.form.get("priority", 5))
-        expiration = int(request.form.get("expiration", 0))
+        priority = float(request.form.get("priority", 5))
+        expiration = float(request.form.get("expiration", 0))
         effects = filter(bool, request.form.get("effects", "").split(","))
         name = request.form.get("name", None)
